+++ Passive.Test/DynamicModelTests/Paged.feature.cs
@@ -17,23 +17,23 @@ namespace Passive.Test.DynamicModelTests
     
     [System.CodeDom.Compiler.GeneratedCodeAttribute("TechTalk.SpecFlow", "1.6.1.0")]
     [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    public partial class DynamicModelSingleFeature : Xunit.IUseFixture<DynamicModelSingleFeature.FixtureData>, System.IDisposable
+    public partial class DynamicModelPagedFeature : Xunit.IUseFixture<DynamicModelPagedFeature.FixtureData>, System.IDisposable
     {
         
         private static TechTalk.SpecFlow.ITestRunner testRunner;
         
-#line 1 "Single.feature"
+#line 1 "Paged.feature"
 #line hidden
         
         public static void FeatureSetup()
         {
             testRunner = TechTalk.SpecFlow.TestRunnerManager.GetTestRunner();
-            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "DynamicModel Single()", "In order to access data\r\nAs a developer\r\nI want to get a single row from the data" +
-                    "base", GenerationTargetLanguage.CSharp, ((string[])(null)));
+            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "DynamicModel Paged()", "In order to reduce the load on my database\r\nAs a developer\r\nI want to get page re" +
+                    "sults", GenerationTargetLanguage.CSharp, ((string[])(null)));
             testRunner.OnFeatureStart(featureInfo);
         }
         
-        public virtual void SetFixture(DynamicModelSingleFeature.FixtureData fixtureData)
+        public virtual void SetFixture(DynamicModelPagedFeature.FixtureData fixtureData)
         {
         }
         
@@ -94,150 +94,192 @@ namespace Passive.Test.DynamicModelTests
         }
         
         [Xunit.FactAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
-        [Xunit.TraitAttribute("Description", "Getting a record without any filtering")]
-        public virtual void GettingARecordWithoutAnyFiltering()
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
+        [Xunit.TraitAttribute("Description", "Getting all records")]
+        public virtual void GettingAllRecords()
         {
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Getting a record without any filtering", ((string[])(null)));
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Getting all records", ((string[])(null)));
 #line 15
 this.ScenarioSetup(scenarioInfo);
 #line 16
   testRunner.Given("I have a model for Appliance");
 #line 17
-  testRunner.When("I ask for a single row");
+  testRunner.When("I ask for page 1");
 #line 18
-  testRunner.Then("I should get 1 result");
+  testRunner.And("the page size is 20");
 #line 19
-  testRunner.And("it should be a subset of all data");
-#line hidden
-            testRunner.CollectScenarioErrors();
-        }
-        
-        [Xunit.FactAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
-        [Xunit.TraitAttribute("Description", "Getting a record by id")]
-        public virtual void GettingARecordById()
-        {
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Getting a record by id", ((string[])(null)));
-#line 21
-this.ScenarioSetup(scenarioInfo);
-#line 22
-  testRunner.Given("I have a model for Appliance");
-#line 23
-  testRunner.When("I ask for the record with the id of 2");
-#line 24
-  testRunner.Then("I should get appliance #2");
-#line hidden
-            testRunner.CollectScenarioErrors();
-        }
-        
-        [Xunit.FactAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
-        [Xunit.TraitAttribute("Description", "Getting a record that doesn\'t exist")]
-        public virtual void GettingARecordThatDoesnTExist()
-        {
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Getting a record that doesn\'t exist", ((string[])(null)));
-#line 26
-this.ScenarioSetup(scenarioInfo);
-#line 27
-  testRunner.Given("I have a model for Appliance");
-#line 28
-  testRunner.When("I ask for the record with the id of 400");
-#line 29
-  testRunner.Then("I should get no results");
+  testRunner.Then("I should get all items");
 #line hidden
             testRunner.CollectScenarioErrors();
         }
         
         [Xunit.Extensions.TheoryAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
+        [Xunit.TraitAttribute("Description", "Getting some records")]
+        [Xunit.Extensions.InlineDataAttribute("1", "2", "2", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("2", "2", "2", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("3", "2", "0", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("1", "3", "3", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("2", "3", "1", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("3", "3", "0", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("1", "4", "4", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("2", "4", "0", new string[0])]
+        public virtual void GettingSomeRecords(string page, string pageSize, string count, string[] exampleTags)
+        {
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Getting some records", exampleTags);
+#line 21
+this.ScenarioSetup(scenarioInfo);
+#line 22
+  testRunner.Given("I have a model for Appliance");
+#line 23
+  testRunner.When(string.Format("I ask for page {0}", page));
+#line 24
+  testRunner.And(string.Format("the page size is {0}", pageSize));
+#line 25
+  testRunner.Then(string.Format("I should get {0} results", count));
+#line 26
+  testRunner.And("they should be a subset of all data");
+#line hidden
+            testRunner.CollectScenarioErrors();
+        }
+        
+        [Xunit.Extensions.TheoryAttribute()]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
         [Xunit.TraitAttribute("Description", "Filtering records by an object")]
         [Xunit.Extensions.InlineDataAttribute("Stainless Steel", "1", new string[0])]
         [Xunit.Extensions.InlineDataAttribute("Red", "1", new string[0])]
-        [Xunit.Extensions.InlineDataAttribute("White", "1", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("White", "2", new string[0])]
         [Xunit.Extensions.InlineDataAttribute("Green", "0", new string[0])]
         public virtual void FilteringRecordsByAnObject(string value, string count, string[] exampleTags)
         {
             TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Filtering records by an object", exampleTags);
-#line 31
+#line 39
 this.ScenarioSetup(scenarioInfo);
-#line 32
+#line 40
   testRunner.Given("I have a model for Appliance");
-#line 33
-  testRunner.When("I ask for a single row");
-#line 34
+#line 41
+  testRunner.When("I ask for page 1");
+#line 42
   testRunner.And(string.Format("I only want appliances colored {0}", value));
-#line 35
-  testRunner.Then(string.Format("I should get {0} results", count));
-#line 36
+#line 43
+  testRunner.Then(string.Format("I should only have {0} results", count));
+#line 44
   testRunner.And(string.Format("I should only get {0}-colored appliances", value));
 #line hidden
             testRunner.CollectScenarioErrors();
         }
         
         [Xunit.Extensions.TheoryAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
         [Xunit.TraitAttribute("Description", "Filtering records by string")]
-        [Xunit.Extensions.InlineDataAttribute("6", "1", new string[0])]
-        [Xunit.Extensions.InlineDataAttribute("7", "1", new string[0])]
-        [Xunit.Extensions.InlineDataAttribute("10", "1", new string[0])]
-        [Xunit.Extensions.InlineDataAttribute("15", "1", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("6", "2", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("7", "2", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("10", "2", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("15", "2", new string[0])]
         [Xunit.Extensions.InlineDataAttribute("20", "1", new string[0])]
         [Xunit.Extensions.InlineDataAttribute("30", "0", new string[0])]
         public virtual void FilteringRecordsByString(string value, string count, string[] exampleTags)
         {
             TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Filtering records by string", exampleTags);
-#line 45
+#line 53
 this.ScenarioSetup(scenarioInfo);
-#line 46
+#line 54
   testRunner.Given("I have a model for Appliance");
-#line 47
-  testRunner.When("I ask for a single row");
-#line 48
+#line 55
+  testRunner.When("I ask for page 1");
+#line 56
+  testRunner.And("the page size is 2");
+#line 57
   testRunner.And(string.Format("I only want appliances with more than {0} amps", value));
-#line 49
-  testRunner.Then(string.Format("I should get {0} results", count));
-#line 50
+#line 58
+  testRunner.Then(string.Format("I should only have {0} results", count));
+#line 59
   testRunner.And(string.Format("I should only get appliances with more than {0} amps", value));
 #line hidden
             testRunner.CollectScenarioErrors();
         }
         
+        [Xunit.Extensions.TheoryAttribute()]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
+        [Xunit.TraitAttribute("Description", "Executing a query with order by")]
+        [Xunit.Extensions.InlineDataAttribute("id", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("name", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("AMPS", new string[0])]
+        public virtual void ExecutingAQueryWithOrderBy(string orderby, string[] exampleTags)
+        {
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Executing a query with order by", exampleTags);
+#line 71
+this.ScenarioSetup(scenarioInfo);
+#line 72
+  testRunner.Given("I have a model for Appliance");
+#line 73
+  testRunner.When("I ask for page 1");
+#line 74
+  testRunner.And(string.Format("I order rows by {0}", orderby));
+#line 75
+  testRunner.Then(string.Format("the records should be sorted by {0}", orderby));
+#line hidden
+            testRunner.CollectScenarioErrors();
+        }
+        
+        [Xunit.Extensions.TheoryAttribute()]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
+        [Xunit.TraitAttribute("Description", "Executing a query with descending order by")]
+        [Xunit.Extensions.InlineDataAttribute("id", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("name", new string[0])]
+        [Xunit.Extensions.InlineDataAttribute("AMPS", new string[0])]
+        public virtual void ExecutingAQueryWithDescendingOrderBy(string orderby, string[] exampleTags)
+        {
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Executing a query with descending order by", exampleTags);
+#line 83
+this.ScenarioSetup(scenarioInfo);
+#line 84
+  testRunner.Given("I have a model for Appliance");
+#line 85
+  testRunner.When("I ask for page 1");
+#line 86
+  testRunner.And(string.Format("I order rows by {0} desc", orderby));
+#line 87
+  testRunner.Then(string.Format("the records should be reverse-sorted by {0}", orderby));
+#line hidden
+            testRunner.CollectScenarioErrors();
+        }
+        
         [Xunit.FactAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
         [Xunit.TraitAttribute("Description", "Selecting a subset of columns")]
         public virtual void SelectingASubsetOfColumns()
         {
             TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Selecting a subset of columns", ((string[])(null)));
-#line 61
+#line 95
 this.ScenarioSetup(scenarioInfo);
-#line 62
+#line 96
   testRunner.Given("I have a model for Appliance");
-#line 63
-  testRunner.When("I ask for a single row");
-#line 64
+#line 97
+  testRunner.When("I ask for page 1");
+#line 98
   testRunner.And("I ask for the columns \"Id, Name\"");
-#line 65
+#line 99
   testRunner.Then("the records should only have the columns \"Id, Name\"");
 #line hidden
             testRunner.CollectScenarioErrors();
         }
         
         [Xunit.FactAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Single()")]
+        [Xunit.TraitAttribute("FeatureTitle", "DynamicModel Paged()")]
         [Xunit.TraitAttribute("Description", "Selecting an invalid column")]
         public virtual void SelectingAnInvalidColumn()
         {
             TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Selecting an invalid column", ((string[])(null)));
-#line 67
+#line 101
 this.ScenarioSetup(scenarioInfo);
-#line 68
+#line 102
   testRunner.Given("I have a model for Appliance");
-#line 69
-  testRunner.When("I ask for a single row");
-#line 70
+#line 103
+  testRunner.When("I ask for page 1");
+#line 104
   testRunner.And("I ask for an invalid column");
-#line 71
+#line 105
   testRunner.Then("the query should throw an exception");
 #line hidden
             testRunner.CollectScenarioErrors();
@@ -250,12 +292,12 @@ this.ScenarioSetup(scenarioInfo);
             
             public FixtureData()
             {
-                DynamicModelSingleFeature.FeatureSetup();
+                DynamicModelPagedFeature.FeatureSetup();
             }
             
             void System.IDisposable.Dispose()
             {
-                DynamicModelSingleFeature.FeatureTearDown();
+                DynamicModelPagedFeature.FeatureTearDown();
             }
         }
     }