+++ CHANGELOG.markdown
@@ -1,6 +1,8 @@
 0.1.3.0
 -------
 * Added missing "args" parameter to DynamicModel.Single() to fully support the where clause.
+* Added DatabaseCapabilities class to track known capabilities.
+* Added paging support for SQL CE 4.0.
 
 0.1.2.0
 -------